add TopDishes component with dynamic loading of top dishes depending on breakpoint
