+++ client/src/pages/Restaurant/Restaurant.js
@@ -6,16 +6,10 @@ import Typography from '@mui/material/Typography';
 import { fixedEncodeURIComponent } from '../../utils/uriEncodeHelpers';
 import RestaurantCard from '../../components/RestaurantCard/RestaurantCard';
 import QuestionsList from '../../components/QuestionsList/QuestionsList';
-import ListItems from '../../components/ListItems/ListItem';
-import useSortedItems from '../../customHooks/useSortedItems';
-import { productsCardWidth } from '../../components/ListItems/styles';
-import ProductCardItem from '../../components/ProductCardItem/ProductCardItem';
+import TopDishes from '../../components/TopDishes/TopDishes';
 
 const RestaurantPage = () => {
-  const products = useSelector((state) => state.products.products, shallowEqual);
-  const sortedProducts = useSortedItems(products, productsCardWidth);
   const partners = useSelector((state) => state.partners.partners, shallowEqual);
-  console.log(partners);
 
   const styleRestaurant = {
     mobile: 315,
@@ -25,35 +19,37 @@ const RestaurantPage = () => {
   };
 
   return (
-    <Container
-      component="section"
-      sx={{
-        bgcolor: 'background.default',
-        my: '60px',
-      }}
-    >
-      <Typography
-        variant="h2"
-        component="h2"
-        color="text.primary"
-        sx={{ textAlign: 'center', mb: 6 }}
+    <>
+      <Container
+        component="section"
+        sx={{
+          bgcolor: 'background.default',
+          my: '60px',
+        }}
       >
-        Our All Restaurants
-      </Typography>
-      {partners.map(({ rating, name, imageUrl, description }) => (
-        <Link key={name} to={`/restaurants/${fixedEncodeURIComponent(name)}`}>
-          <RestaurantCard
-            rating={rating}
-            name={name}
-            imageUrl={imageUrl}
-            description={description}
-            styleWidth={styleRestaurant}
-          />
-        </Link>
-      ))}
-      <ListItems title="Our Top Dishes" topDish items={sortedProducts} itemComponent={ProductCardItem} actions={null} />
+        <Typography
+          variant="h2"
+          component="h2"
+          color="text.primary"
+          sx={{ textAlign: 'center', mb: 6 }}
+        >
+          Our All Restaurants
+        </Typography>
+        {partners.map(({ rating, name, imageUrl, description }) => (
+          <Link key={name} to={`/restaurants/${fixedEncodeURIComponent(name)}`}>
+            <RestaurantCard
+              rating={rating}
+              name={name}
+              imageUrl={imageUrl}
+              description={description}
+              styleWidth={styleRestaurant}
+            />
+          </Link>
+        ))}
+      </Container>
+      <TopDishes />
       <QuestionsList />
-    </Container>
+    </>
   );
 };
